requirements.txt
cmake==3.18.0
dlib==19.20.0
imutils==0.5.3
numpy==1.19.0
opencv-contrib-python==4.3.0.36
scipy==1.5.1
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
